requirements.txt
fastapi
uvicorn
pydantic
spacy
phonenumbers
python-dateutil
